Stabilize two-LUN startup and flexible storage
Cache payload inventories so drive snapshots no longer stall boot, serialize Ofrenda host/local access, and import only after clean ejects. Remove fixed content-count limits, add storage reclamation coverage, and update the portal, TUI, documentation, and host tests.

+++ src/usb_drive.h
@@ -10,8 +10,8 @@ struct UsbDriveState {
   // the rings are walked.
   bool mediaPresent;
   uint16_t noteCount;
-  uint8_t scriptCount;
-  uint8_t badusbScriptCount;
+  uint16_t scriptCount;
+  uint16_t badusbScriptCount;
   uint16_t tagCount;
 };
